% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_report.R
\name{get_food_report}
\alias{get_food_report}
\title{Retrieve food reports from the USDA Food Compositions  Data Base.}
\source{
\url{https://ndb.nal.usda.gov/ndb/doc/apilist/API-FOOD-REPORTV2.md}
}
\usage{
get_food_report(api_key = NULL, ndbno = NULL, type = "b",
  result_type = "json")
}
\arguments{
\item{api_key}{A string. A unique api_key obtained after signingup at [NDB](https://ndb.nal.usda.gov/ndb/doc/index#.) If this is not set with
\code{\link{set_apikey}}, then the user can manually set it here.}

\item{ndbno}{A string. A unique ndb number for a specific food of interest. Use this if nutrients is set
to `NULL`.}

\item{type}{One of "b", "f" or "s" for basic, full or stats}

\item{result_type}{A string. One of json or xml depending on  the type of format required.}
}
\value{
An object as per the requested format.
}
\description{
Retrieve food reports from the USDA Food Compositions  Data Base.
}
\details{
This uses Food Reports API Version 2
}
\examples{
\dontrun{
 get_food_report(ndbno = "35193",
result_type = "xml")
get_food_report(ndbno = c("35193","01009"),
               result_type = "xml")
               res<-get_food_report(ndbno = "35193",
               result_type = "xml")
               pretty_xml(res, tag="food", target="type")
               }
}
\seealso{
\code{\link{get_nutrients}} \code{\link{pretty_xml}} \code{\link{pretty_json}}
}
