% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{get_list}
\alias{get_list}
\title{Get Lists from the USDA Food Composition Databases}
\usage{
get_list(list_type = "f", sort_by = "n", api_key = NULL,
  max_items = 30, offset = 0, format = "json",
  return_request = FALSE)
}
\arguments{
\item{list_type}{A character(string). One of f,n,ns,nr,f, and g. See details
below.}

\item{sort_by}{A character(string). How would you like to sort the lists? One of 
n for name and id. id differs depending on the list type.}

\item{api_key}{string. A valid api_key obtained from \url{https://ndb.nal.usda.gov/ndb/doc/index#}}

\item{max_items}{integer. Maximum number of items to return.}

\item{offset}{integer. Where should querying start? Defaults to 0.}

\item{format}{string. What kind of output do you need? Supports "xml" and "json"}

\item{return_request}{If format is set to json, this controls whether the initial request 
is returned to the user. Defaults to FALSE.}
}
\value{
A list depending on the requested format type. For JSON, a list of length 2 if return request is requested. Otherwise a 
list of length 1 with a semi_processed JSON object. The return value for XML is 
an xml document.
}
\description{
A convenient wrapper function to obtain lists from the 
USDA nutrient databases. It queries the database for a list of foods,
food groups or nutrients.
}
\details{
For list_type d represents derived foods, f for food,
n for all nutrients, ns speciality nutrients, nr for standard nutrients 
and g for food group. Details available at \url{https://ndb.nal.usda.gov/ndb/doc/apilist/API-LIST.md}
}
\examples{
\dontrun{
res <- get_list(
list_type = "ns",sort_by = "id",
max_items = 50,offset = 12,format = "xml")
res
usdar::pretty_xml(res,tag="name")
usdar::pretty_json(res)
}
}
\references{
\url{https://ndb.nal.usda.gov/ndb/doc/apilist/API-LIST.md}
}
\seealso{
\code{\link{get_nutrients}} \code{\link{pretty_xml}} \code{\link{pretty_json}}
}
