% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrient_id.R
\name{get_nutrient_id}
\alias{get_nutrient_id}
\title{Convenience function to easily extract nutrient ids.}
\usage{
get_nutrient_id(nutrient_name, df)
}
\arguments{
\item{nutrient_name}{Name of the nutrient for which an id is returned.}

\item{df}{The data set that contains nutrient ids. Can be loaded into the 
environment by calling `data("nutrient_ids")`. If no data set is provided, the default is to 
use `nutrient_ids`.}
}
\value{
Returns the nutrient id.
}
\description{
Provides an easy way to access nutrient ids as provided 
by the USDA.
}
\examples{
data("nutrient_ids")
get_nutrient_id("myri",nutrient_ids)
}
