% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulators.R
\name{get_nutrient_info}
\alias{get_nutrient_info}
\title{Convert JSON data to a simpler data.frame object.}
\usage{
get_nutrient_info(result)
}
\arguments{
\item{result}{An object obtained from `get_nutrients`}
}
\description{
When JSON data is requested, `get_nutrient info` provides
a convenient way to get semi-clean data.
}
