% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nutrients.R
\name{get_nutrients}
\alias{get_nutrients}
\title{USDA Nutrient Database Information Retrieval.}
\source{
\url{\link{https://ndb.nal.usda.gov/ndb/doc/apilist/API-NUTRIENT-REPORT.md}}
}
\usage{
get_nutrients(result_type = "json", nutrients = NULL, api_key = NULL,
  ndbno = NULL, subset = 0, offset = 0, max_rows = 50,
  food_group = NULL)
}
\arguments{
\item{result_type}{A string. One of json or xml depending on  the type of format required.}

\item{nutrients}{A string or vector of strings. A nutrient id as provided on the nutrient data base website.}

\item{api_key}{A string. A unique api_key obtained after signingup at [NDB](https://ndb.nal.usda.gov/ndb/doc/index#.) If this is not set with
\code{\link{set_apikey}}, then the user can manually set it here.}

\item{ndbno}{A string. A unique ndb number for a specific food of interest. Use this if nutrients is set
to `NULL`.}

\item{subset}{Numeric. Defaults to 0 for all food types. 1 to return common food types.}

\item{offset}{Where should offsetting begin? Defaults to 0. Lower values mean more results.}

\item{max_rows}{Numeric. Maximum number of rows to return. Defaults to 50.}

\item{food_group}{String. Return results for only a certain food group.}
}
\value{
A list of length two. The first list element returns unprocessed data if JSON is requested while the second is a data.frame object from
JSON data. An XML file is returned if xml is requested.
}
\description{
A simple function to access and retrieve nutrient data
from the USDA nutrient data base.
}
\examples{
\dontrun{
get_nutrients(result_type = "json", nutrients = "204",
food_group = "0500",
max_rows = 50, 
subset = 1)
}
}
\references{
U.S. Department of Agriculture, Agricultural Research Service. 20xx. USDA National Nutrient Database for Standard Reference, Release . Nutrient Data Laboratory Home Page, http://www.ars.usda.gov/nutrientdata
}
