% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_info.R
\name{get_report_info}
\alias{get_report_info}
\title{Extraction of data from a food report}
\usage{
get_report_info(what, result)
}
\arguments{
\item{what}{A `character`(string). What kind of information do you want to extract?!}

\item{result}{An object obtained from `get_nutrients`}
}
\value{
A list containing requested report info.
}
\description{
This function provides a quick and efficient way to extract
any form of data about a food report.
}
\examples{
\dontrun{
get_report_info("name",result)
}
}
\seealso{
\code{\link{get_nutrient_info}}
\code{\link{get_nutrients}}
}
