% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_ids.R
\docType{data}
\name{nutrient_ids}
\alias{nutrient_ids}
\title{A Curated list of nutrient IDs linked to Pubmed data.
This data set contains nutrient_ids as provided by the USDA NDB.}
\format{'data.frame':	125 obs. of  4 variables:
\describe{
\item{PubChem.CID.s.}{Integer code corresponding to a PubChem ID}
\item{NDB.ID}{The corresponding NDB nutrient id}
\item{NDB.Name}{Nutrient's common name}

}}
\source{
\url{https://bmcresnotes.biomedcentral.com/articles/10.1186/s13104-018-3997-y}
}
\usage{
nutrient_ids
}
\description{
A Curated list of nutrient IDs linked to Pubmed data.
This data set contains nutrient_ids as provided by the USDA NDB.
}
\keyword{datasets}
