% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_ids_clean.R
\docType{data}
\name{nutrient_ids_clean}
\alias{nutrient_ids_clean}
\title{Extended Nutrient ID data.
This extends the ids in `nutrient_ids` as provided by the USDA NDB.}
\format{'data.frame':	234 obs. of  2 variables:
\describe{
\item{NDB.ID}{The corresponding NDB nutrient id}
\item{NDB.Name}{Nutrient's common name}

}}
\source{
\url{https://bmcresnotes.biomedcentral.com/articles/10.1186/s13104-018-3997-y}
}
\usage{
nutrient_ids_clean
}
\description{
Extended Nutrient ID data.
This extends the ids in `nutrient_ids` as provided by the USDA NDB.
}
\keyword{datasets}
