% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_xml.R
\name{pretty_xml}
\alias{pretty_xml}
\title{Process XML data}
\usage{
pretty_xml(result, tag, target = NULL, return_raw = FALSE)
}
\arguments{
\item{result}{An object obtained from get_nutrients with
result_type set to xml}

\item{tag}{An xml tag.}

\item{target}{Character. Extract certain attributes from processed xml data. Defaults to 
NULL.}

\item{return_raw}{Logical. Do you want to get a raw xml file and proceed with processing yourself? Defaults
to FALSE.}
}
\value{
If default parameters are used, a semi processed xml file is returned.
If target is set, xml attributes corrseponding to target are returned.
}
\description{
This function is useful if one has requested for the return output
to be of xml type in get_nutrients.
}
\examples{
\dontrun{
if(!require(xml2)) require(xml2)
pretty_xml(res,"food")
}
}
