% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulators.R
\name{get_nutrient_info}
\alias{get_nutrient_info}
\title{Convert JSON data to a simpler data.frame object.}
\usage{
get_nutrient_info(res, bind_data = TRUE, abbr = TRUE,
  abbr_limit = 14)
}
\arguments{
\item{res}{An object of type `json` obtained from `get_nutrients`}

\item{bind_data}{Logical. Should the data be bound as a single `data.frame` object? Defaults
to TRUE.}

\item{abbr}{Logical. Should the names of Source be abbrevaited? Defaults to  TRUE.}

\item{abbr_limit}{If abbr is TRUE, this determines the abbreviation limit. Defaults to 
14.}
}
\value{
The default return value is a data.frame object showing
the corresponding data as requested with `get_nutrients`. If bind_data is
set to FALSE, a list of data frames is returned instead.
}
\description{
When JSON data is requested, `get_nutrient info` provides
a convenient way to get semi-clean data.
}
\examples{
\dontrun{
res<-get_nutrients(nutrients = "204")
head(get_nutrient_info(res))
}
}
\seealso{
\code{\link{get_nutrients}}
}
