% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulators.R
\name{pretty_json}
\alias{pretty_json}
\title{Pretty JSON Lists}
\usage{
pretty_json(result)
}
\arguments{
\item{result}{An object obtained from `get_nutrients`}
}
\value{
A pretty `dput` style list.
}
\description{
A convenience function that returns processed JSON data.
}
